Replace vars from empty inner side with NULL constants
If the inner side of an outer join is empty, the result is equivalent to
the projecting targetlist after replacing the inner vars with NULL
constants. This commit does the same during set_join_references(). This
does not show any performance improvement as of its own. A detailed fix
is on its way.

Ashutosh Bapat