Report memory used for planning a query in EXPLAIN
The memory used is reported as "Planning Memory" property in EXPLAIN output
when any of options ANALYZE or SUMMARY is specified.

The memory used in the CurrentMemoryContext and its children is noted before
and after calling pg_plan_query() from ExplainOneQuery(). The difference in the
two values is reported as the memory consumed while planning the query. This
may not account for the memory allocated in memory contexts which are not
children of CurrentMemoryContext when calling pg_plan_query(). These contexts
are usually other long lived contexts, e.g.  CacheMemoryContext, which are
shared by all the queries run in a session. The consumption in those can not be
attributed only to a given query and hence should not be reported any way.

Ashutosh Bapat