Use RestrictInfo hash table for storing EC derived child RestrictInfos
RestrictInfos derived from an EquivalenceClass, including the child
RestrictInfos, are stored in its EquivalenceClass::ec_derives which
is a linked list. The linked list is scanned every time
create_join_clause() is called, which in turn is called multiple times
for a given partition. The loop thus contributes O(N^2) at run time
where N is number of partitions. When there are thousands of
partition, the loops contributes significantly to the planning time.

In order to reduce the time complexity of searching a clause, store the
child clauses in a hash table, so that contribution of this loop in
overall planning time reduces to O(N). Please note that the parent
derived clauses are still stored in EquivalenceClass::ec_derives. The
hash table approach differs from the linked list based approach in the
following aspects:

1. The linked lists are per EC whereas we use a single hash table for
   all the child clauses. Linked list requires less storage compared to
   a hash table, thus having one hash table per EquivalencClass consumes
   more memory. Hash tables are more efficient thus having one to store
   clauses from all the ECs doesn't affect time efficiency much.

2. EquivalenceClass::ec_derives is searched by
   EquivalenceClass::left_em/right_em pointers whereas the hash table is
   key'ed by RestrictInfo::rinfo_serial and
   RestrictInfo::required_relids. In the future (next commits) the hash
   table will be used for storing the child RestrictInfos translated
   from parent RestrictInfos which are not part of ECs as well. For
   RestrictInfos which are not part of ECs,
   EquivalenceClass::left_em/right_em are not set and hence can not be
   used as a key. OTOH, RestrictInfo::rinfo_serial and
   RestrictInfo::required_relids is set for all RestrictInfos.
   create_join_clause() searches for the parent clause in
   EquivalenceClass::ec_sources or EquivalenceClass::ec_derives before
   searching for the child clause in the hash table. That's an extra
   work compared to linked list based approach where a child clause
   could be searched directly in EquivalenceClass::ec_derives. But with
   this change, the linked lists should be short and thus searching for
   a parent clause should not take much time. If the child clause needs
   to be created, the parent clause is anyway needed to copy
   RestrictInfo::rinfo_serial from parent clause into the child clause
   so nothing extra happens there.

Please note that EquivalenceClass:ec_derives is used in functions other
than create_join_clause(). But all those usages happen before any child
EC members appear. There are Asserts in those places to validate this
fact. Also the Asserts serve a notice in case things change in future.
Hence moving child derived clauses out of EquivalenceClass does not need
changes to those functions.

Ashutosh Bapat