Avoid translating RestrictInfo repeatedly
RestrictInfo for the child relations (including the child join
relations) are obtained by translating RestrictInfo applicable to the
parent rel. Since these translations are not tracked, the same
RestrictInfo may get translated multiple times for the same parent and
child pair. When using partitionwise join this can happen as many times
as the number of possible join orders between the partitioned tables and
as many times a parent path is reparameterized for a child if a clause
is used in such a path.

Repeated translations are avoided by saving them in a hash table hanging
off of PlannerInfo.  RestrictInfo::rinfo_serial and
RestrictInfo::required_relids are used to compute hash key since they
together identify a RestrictInfo uniquely.  Two functions
find_child_rinfo() and add_child_rinfo() are used to search a
RestrictInfo for a given tanslation and add a child RestrictInfo
respectively.

Ashutosh Bapat