\d variants for property graphs
Since property graphs objects do not have a predefined set of columns
like other relations, we can not print columns and type for them as we
do for other relations. Going by the example of sequences, instead we
print things that define the state of property graph, which is its
element.  Hence \d of a property graph shows the list of elements, their
underlying tables, element kinds and end vertex aliases for elements of
kind edge.  \d+ shows property graph definition which contains labels
and properties as well.

\dG list the property graph objects with names matching the given
pattern. \dG+ addionally prints owner and the description.

Additionally \? help now mentions property graphs whever applicable.

Notes to reviewers:
1. It seems waste of cycles to pg_strdup and constant string "Property
   graph definition" and then free it. But that's what listSchemas()
seems to be doing for printing "Publications" footer. So did it here as
well.

2. If you wonder why g1 and g2 are both \d and \d+ 'ed, g1 is an empty
   property graph and g2 is a non-emptry one. I think it's apparent from
the output so didn't add a comment. But in case it's worth commenting we
can add a comment.

Author: Ashutosh Bapat
Author: Junwang Zhou